# proto/xlpod.openapi.yaml
openapi: 3.1.0
info:
  title: xlpod
  version: 0.0.0
  summary: Loopback HTTPS API for the xlpod launcher.
  description: |
    Single source of truth for the xlpod loopback HTTPS API. Both the Rust
    launcher (`launcher/`) and the pure-python client (`client/xlpod`, future)
    are derived from and validated against this file. Neither side imports the
    other; this spec is the only contact surface.

    See `proto/README.md` for the change procedure and `docs/SECURITY.md` for
    the threat model that pins every constraint declared here.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  contact:
    name: xlpod maintainers
    url: https://github.com/eddmpython/xlpod

  # ---- xlpod-specific extensions (machine-readable; the launcher reads these
  # at compile time and refuses to build if any constraint is relaxed). ----
  x-xlpod-proto: 1
  x-xlpod-bind:
    - "127.0.0.1"
    - "::1"
  x-xlpod-tls-only: true
  x-xlpod-allowed-origins:
    - "https://addin.xlwings.org"
  x-xlpod-rate-limit:
    per_token_rps: 100
  x-xlpod-audit: true
  x-xlpod-notes: |
    - 0.0.0.0 binding is forbidden at compile time. Loopback only.
    - Plain HTTP is forbidden. xlwings Lite enforces upgrade-insecure-requests
      and block-all-mixed-content (Phase 0 measurement, 2026-04-07), so a
      plain-text fallback would be unreachable from the only legitimate
      caller and is therefore deleted to shrink attack surface.
    - The Origin header MUST exactly match an entry in
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/handshake:
    post:
      operationId: postAuthHandshake
      tags: [auth]
      summary: Issue a fresh bearer token with a narrow scope set.
      description: |
        Validates the Origin header against `x-xlpod-allowed-origins`,
        canonicalizes any requested `fs_roots`, then **prompts the user
        for consent via the launcher's tray dialog** before issuing a
        token. If the user denies (or the consent backend rejects), the
        call returns 403 with code `consent_denied` and no token is
        ever created. On approval the launcher mints a new 256-bit
        random bearer token bound to the granted scopes and roots.

        The user's decision is the final word: a launcher with the
        default `MessageBoxConsent` backend cannot be bypassed from the
        wire. Test and dev binaries can replace the backend with
        `AutoApproveConsent`; the production tray launcher does not
        expose that escape.

        Tokens are NEVER persisted to disk and NEVER logged (the audit log
        records only the token id, not the secret). A new token is issued at
        every launcher start; restarting the launcher invalidates all
        outstanding tokens.
      security: []
      parameters:
        - $ref: "#/components/parameters/XXLPodProto"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandshakeRequest"
      responses:
        "200":
          description: Token issued (possibly with a narrowed scope set).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Handshake"
        "403":
          description: Origin not in allow-list, or user denied consent.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /fs/read:
    get:
      operationId: getFsRead
      tags: [fs]
      summary: Read a file under one of the token's approved fs roots.
      description: |
        Reads a single file from disk and returns its content
        base64-encoded. The path is resolved against the launcher's
        current working directory if relative, then canonicalized
        (symlinks followed). The canonical path MUST start with one of
        the `fs_roots` the token was issued with — paths outside the
        approved set are rejected with `forbidden_path`, never read.

        Size is capped at 10 MiB; oversized files are rejected with
        `path_too_large` *before* the read is performed. Directories,
        devices, and missing files are rejected with `not_a_file` /
        `path_not_found`. The handler does not follow into directories
        and does not list — listing is a separate route added with
        `/fs/list` in a later phase.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XXLPodProto"
        - name: path
          in: query
          required: true
          description: Absolute or relative path. Symlinks are followed; the resolved path must lie under an approved root.
          schema:
            type: string
      responses:
        "200":
          description: File contents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "400":
          description: Malformed path or oversized file.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403":
          description: Token lacks `fs:read`, or path is outside the approved roots.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "404":
          description: Path does not exist or is not a file.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /run/python:
    post:
      operationId: postRunPython
      tags: [run]
      summary: Execute a Python snippet inside the launcher's worker.
      description: |
        Sends `code` to a long-running Python worker process owned by
        the launcher and returns its captured stdout/stderr plus an
        optional `result` (the value of the `_result` global if the
        snippet sets one). The worker is shared across all callers
        within a single launcher run; calls are serialized so two
        snippets never execute concurrently against the same globals
        namespace.

        **Execution errors are NOT HTTP errors.** A snippet that
        raises (`1/0`, ImportError, etc.) returns 200 with `ok=false`
        and a Python traceback in `error`. The HTTP layer only fails
        when the worker itself misbehaves: `worker_spawn_failed` if
        the launcher cannot find a Python interpreter,
        `worker_timeout` if the snippet exceeds the wall-clock cap
        (default 30 s; the worker is killed and the next call gets a
        fresh process), or `worker_crashed` if the worker dies
        mid-call.

        Phase 5 ships a single shared worker. Per-workbook isolation,
        idle shutdown, and the 800 MiB memory ceiling described in
        `docs/design.md` §3.6 land in Phase 5.x.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XXLPodProto"
        - $ref: "#/components/parameters/XXLPodPlanOnly"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunPythonRequest"
      responses:
        "200":
          description: Worker executed (possibly with `ok=false` for a Python-level error).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunPythonResponse"
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403":
          description: Token lacks `run:python`.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "500":
          description: Worker spawn failed or worker crashed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "504":
          description: Worker exceeded the wall-clock cap; the worker has been killed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /excel/workbooks:
    get:
      operationId: getExcelWorkbooks
      tags: [excel]
      summary: List workbooks open in the running Excel instance.
      description: |
        Attaches to the user's already-running Excel via COM
        (`GetActiveObject("Excel.Application")`) and returns the
        currently open workbooks. Does **not** start a new Excel
        instance — Phase 6 is read-only against whatever Excel the
        user already opened.

        Requires `excel:com` scope. Returns 503 with
        `excel_not_available` if `pywin32` is missing from the
        worker's Python (e.g. on a Linux CI runner), or
        `excel_not_running` if Excel is not currently open.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XXLPodProto"
      responses:
        "200":
          description: List of open workbooks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkbookList"
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403":
          description: Token lacks `excel:com`.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "503":
          description: Excel COM unavailable on the worker.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /excel/range/read:
    post:
      operationId: postExcelRangeRead
      tags: [excel]
      summary: Read a range from a workbook in the running Excel instance.
      description: |
        Reads `range` from `sheet` in `workbook` (all three are
        Excel-side names; the workbook must already be open). The
        result is a 2-D array of cell values; single cells come back
        as `[[value]]` so the response shape is uniform. Empty cells
        return `null`. The canonical Excel address Excel resolved to
        is returned in `address` for round-trip clarity.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XXLPodProto"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RangeReadRequest"
      responses:
        "200":
          description: Range values.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RangeData"
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403":
          description: Token lacks `excel:com`.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "503":
          description: Excel COM unavailable on the worker.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /ai/chat:
    post:
      operationId: postAiChat
      tags: [ai]
      summary: Non-streaming chat completion against an AI provider.
      description: |
        Relays a chat completion request to the configured AI provider
        (Phase 8 ships Anthropic only). The launcher loads the API key
        from the OS keychain — the key never appears in the request,
        the response, or any audit line. The model receives the
        launcher's tool registry (see `GET /ai/tools`); when it emits
        a `tool_use` block the launcher dispatches the call back
        through its own routes, traversing the full 5-check stack
        with an internal bearer minted at `POST /ai/session` time.

        Mutating tool calls trigger the consent dialog *per call* in
        Phase 8. Trust windows that skip the dialog for N minutes are
        added in Phase 9. Streaming (`/ai/chat/stream`) is also
        Phase 9.

        Requires `ai:provider:call`. Mutating tools additionally
        require their own scope (`fs:write`, `ai:exec:python`, etc.)
        on the *user's* token at session-open time; the launcher
        bakes the intersection into the AI's internal bearer.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XXLPodProto"
        - $ref: "#/components/parameters/XXLPodPlanOnly"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat completion (possibly after multiple internal tool turns).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403":
          description: Missing scope, or consent denied for a tool call.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "412":
          description: Provider not configured (no API key in keychain).
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "502":
          description: Upstream provider returned an error.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
...
  /ai/providers:
    get:
      operationId: getAiProviders
      tags: [ai]
      summary: List configured AI providers.
      description: |
        Returns each known provider with a *boolean* indicating
        whether a key is present in the OS keychain. The actual key
        is never echoed.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XXLPodProto"
      responses:
        "200":
          description: Provider configuration status.
          content:
            application/json:
              schema:
                type: object
                required: [providers]
                properties:
                  providers:
                    type: array
                    items: {$ref: "#/components/schemas/Provider"}
        "401": {$ref: "#/components/responses/Unauthorized"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-256bit
      description: |
        256-bit random token, generated from the OS CSPRNG at every launcher
        start. Required on every route except `/health`. Tokens are scoped
        (see `Scope`) and rate-limited (100 req/s/token).
...